[FLINK-34209] Migrate FileSink to the new SinkV2 API
