+++ flink-connectors/flink-connector-files/src/main/java/org/apache/flink/connector/file/sink/FileSink.java
@@ -26,9 +26,12 @@ import org.apache.flink.api.common.serialization.BulkWriter;
 import org.apache.flink.api.common.serialization.Encoder;
 import org.apache.flink.api.common.typeinfo.TypeInformation;
 import org.apache.flink.api.connector.sink2.Committer;
-import org.apache.flink.api.connector.sink2.StatefulSink;
-import org.apache.flink.api.connector.sink2.StatefulSink.WithCompatibleState;
-import org.apache.flink.api.connector.sink2.TwoPhaseCommittingSink;
+import org.apache.flink.api.connector.sink2.CommitterInitContext;
+import org.apache.flink.api.connector.sink2.Sink;
+import org.apache.flink.api.connector.sink2.SinkWriter;
+import org.apache.flink.api.connector.sink2.SupportsCommitter;
+import org.apache.flink.api.connector.sink2.SupportsWriterState;
+import org.apache.flink.api.connector.sink2.WriterInitContext;
 import org.apache.flink.api.java.typeutils.EitherTypeInfo;
 import org.apache.flink.connector.file.sink.committer.FileCommitter;
 import org.apache.flink.connector.file.sink.compactor.FileCompactStrategy;
@@ -48,7 +51,7 @@ import org.apache.flink.core.fs.FileSystem;
 import org.apache.flink.core.fs.Path;
 import org.apache.flink.core.io.SimpleVersionedSerializer;
 import org.apache.flink.streaming.api.connector.sink2.CommittableMessage;
-import org.apache.flink.streaming.api.connector.sink2.WithPreCommitTopology;
+import org.apache.flink.streaming.api.connector.sink2.SupportsPreCommitTopology;
 import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.streaming.api.datastream.SingleOutputStreamOperator;
 import org.apache.flink.streaming.api.functions.sink.filesystem.BucketAssigner;
@@ -128,10 +131,11 @@ import static org.apache.flink.util.Preconditions.checkState;
  */
 @Experimental
 public class FileSink<IN>
-        implements StatefulSink<IN, FileWriterBucketState>,
-                TwoPhaseCommittingSink<IN, FileSinkCommittable>,
-                WithCompatibleState,
-                WithPreCommitTopology<IN, FileSinkCommittable>,
+        implements Sink<IN>,
+                SupportsWriterState<IN, FileWriterBucketState>,
+                SupportsCommitter<FileSinkCommittable>,
+                SupportsWriterState.WithCompatibleState,
+                SupportsPreCommitTopology<FileSinkCommittable, FileSinkCommittable>,
                 SupportsConcurrentExecutionAttempts {
 
     private final BucketsBuilder<IN, ? extends BucketsBuilder<IN, ?>> bucketsBuilder;
@@ -141,13 +145,18 @@ public class FileSink<IN>
     }
 
     @Override
-    public FileWriter<IN> createWriter(InitContext context) throws IOException {
+    public SinkWriter<IN> createWriter(InitContext context) throws IOException {
+        throw new UnsupportedOperationException("Not supported");
+    }
+
+    @Override
+    public FileWriter<IN> createWriter(WriterInitContext context) throws IOException {
         return restoreWriter(context, Collections.emptyList());
     }
 
     @Override
     public FileWriter<IN> restoreWriter(
-            InitContext context, Collection<FileWriterBucketState> recoveredState)
+            WriterInitContext context, Collection<FileWriterBucketState> recoveredState)
             throws IOException {
         FileWriter<IN> writer = bucketsBuilder.createWriter(context);
         writer.initializeState(recoveredState);
@@ -167,7 +176,8 @@ public class FileSink<IN>
     }
 
     @Override
-    public Committer<FileSinkCommittable> createCommitter() throws IOException {
+    public Committer<FileSinkCommittable> createCommitter(CommitterInitContext context)
+            throws IOException {
         return bucketsBuilder.createCommitter();
     }
 
@@ -181,6 +191,11 @@ public class FileSink<IN>
             // IOException.
             throw new FlinkRuntimeException("Could not create committable serializer.", e);
         }
+    }
+
+    @Override
+    public SimpleVersionedSerializer<FileSinkCommittable> getWriteResultSerializer() {
+        return getCommittableSerializer();
     }
 
     @Override
@@ -291,7 +306,7 @@ public class FileSink<IN>
         }
 
         @Internal
-        abstract FileWriter<IN> createWriter(final InitContext context) throws IOException;
+        abstract FileWriter<IN> createWriter(final WriterInitContext context) throws IOException;
 
         @Internal
         abstract FileCommitter createCommitter() throws IOException;
@@ -409,7 +424,7 @@ public class FileSink<IN>
         }
 
         @Override
-        FileWriter<IN> createWriter(InitContext context) throws IOException {
+        FileWriter<IN> createWriter(WriterInitContext context) throws IOException {
             OutputFileConfig writerFileConfig;
             if (compactStrategy == null) {
                 writerFileConfig = outputFileConfig;
@@ -603,7 +618,7 @@ public class FileSink<IN>
         }
 
         @Override
-        FileWriter<IN> createWriter(InitContext context) throws IOException {
+        FileWriter<IN> createWriter(WriterInitContext context) throws IOException {
             OutputFileConfig writerFileConfig;
             if (compactStrategy == null) {
                 writerFileConfig = outputFileConfig;
