+++ flink-connectors/flink-connector-files/src/main/java/org/apache/flink/connector/file/sink/writer/FileWriter.java
@@ -21,9 +21,9 @@ package org.apache.flink.connector.file.sink.writer;
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.annotation.VisibleForTesting;
 import org.apache.flink.api.common.operators.ProcessingTimeService;
+import org.apache.flink.api.connector.sink2.CommittingSinkWriter;
 import org.apache.flink.api.connector.sink2.SinkWriter;
-import org.apache.flink.api.connector.sink2.StatefulSink.StatefulSinkWriter;
-import org.apache.flink.api.connector.sink2.TwoPhaseCommittingSink;
+import org.apache.flink.api.connector.sink2.StatefulSinkWriter;
 import org.apache.flink.connector.file.sink.FileSink;
 import org.apache.flink.connector.file.sink.FileSinkCommittable;
 import org.apache.flink.core.fs.Path;
@@ -62,7 +62,7 @@ import static org.apache.flink.util.Preconditions.checkState;
 @Internal
 public class FileWriter<IN>
         implements StatefulSinkWriter<IN, FileWriterBucketState>,
-                TwoPhaseCommittingSink.PrecommittingSinkWriter<IN, FileSinkCommittable>,
+                CommittingSinkWriter<IN, FileSinkCommittable>,
                 SinkWriter<IN>,
                 ProcessingTimeService.ProcessingTimeCallback {
