+++ flink-streaming-java/src/test/java/org/apache/flink/streaming/util/TestExpandingSinkDeprecated.java
@@ -21,26 +21,28 @@ package org.apache.flink.streaming.util;
 import org.apache.flink.api.connector.sink2.Committer;
 import org.apache.flink.api.connector.sink2.CommitterInitContext;
 import org.apache.flink.api.connector.sink2.Sink;
-import org.apache.flink.api.connector.sink2.SinkWriter;
-import org.apache.flink.api.connector.sink2.SupportsCommitter;
-import org.apache.flink.api.connector.sink2.WriterInitContext;
 import org.apache.flink.core.io.SimpleVersionedSerializer;
 import org.apache.flink.streaming.api.connector.sink2.CommittableMessage;
-import org.apache.flink.streaming.api.connector.sink2.SupportsPostCommitTopology;
-import org.apache.flink.streaming.api.connector.sink2.SupportsPreCommitTopology;
-import org.apache.flink.streaming.api.connector.sink2.SupportsPreWriteTopology;
+import org.apache.flink.streaming.api.connector.sink2.WithPostCommitTopology;
+import org.apache.flink.streaming.api.connector.sink2.WithPreCommitTopology;
+import org.apache.flink.streaming.api.connector.sink2.WithPreWriteTopology;
 import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.streaming.api.functions.sink.v2.DiscardingSink;
 
 import java.io.IOException;
 
-/** A test sink that expands into a simple subgraph. Do not use in runtime. */
-public class TestExpandingSinkWithMixin
+/**
+ * A test sink that expands into a simple subgraph. Do not use in runtime.
+ *
+ * <p>Should be removed along with {@link
+ * org.apache.flink.api.connector.sink2.TwoPhaseCommittingSink}
+ */
+@Deprecated
+public class TestExpandingSinkDeprecated
         implements Sink<Integer>,
-                SupportsCommitter<Integer>,
-                SupportsPreWriteTopology<Integer>,
-                SupportsPreCommitTopology<Integer, Integer>,
-                SupportsPostCommitTopology<Integer> {
+                WithPreWriteTopology<Integer>,
+                WithPreCommitTopology<Integer, Integer>,
+                WithPostCommitTopology<Integer, Integer> {
 
     @Override
     public void addPostCommitTopology(DataStream<CommittableMessage<Integer>> committables) {
@@ -59,12 +61,8 @@ public class TestExpandingSinkWithMixin
     }
 
     @Override
-    public SinkWriter<Integer> createWriter(WriterInitContext context) throws IOException {
-        return null;
-    }
-
-    @Override
-    public SinkWriter<Integer> createWriter(InitContext context) throws IOException {
+    public PrecommittingSinkWriter<Integer, Integer> createWriter(InitContext context)
+            throws IOException {
         return null;
     }
 
@@ -75,11 +73,6 @@ public class TestExpandingSinkWithMixin
 
     @Override
     public SimpleVersionedSerializer<Integer> getCommittableSerializer() {
-        return null;
-    }
-
-    @Override
-    public SimpleVersionedSerializer<Integer> getWriteResultSerializer() {
         return null;
     }
 }
